unused: fully deduplicate objects
Previously, we wouldn't actually deduplicate objects, but simply
ignore objects which were known to be duplicated, where the duplicate
was marked as used. This, however, would not subsequently mark other
objects used. Specifically, for the two subgraphs 'A1 -> B' and 'Root
-> A2', A2 would be used, causing A1 to be supressed, but we would
still report B.

Now we correctly deduplicate nodes, so that only ever one node 'A'
exists, creating the graph 'Root -> A -> B'. This does require us to
keep a single graph for all packages even when not operating in whole
program mode. However, that makes sense: even the normal mode of
operation is, on a smaller scale, a whole program analysis, because a
single source package may exist as many different packages in the
build system. The difference between whole-program mode and normal
mode, now, is only which objects get automatically marked as used.

We will need to restore proper parallelism in a later commit.